requirements.txt
tiktoken
chromadb
langchain
langchain-community
langchain-openai
langchain_chroma
gradio
python-dotenv